requirements.txt
amqp==1.4.6
anyjson==0.3.3
billiard==3.3.0.20
celery==3.1.18
cffi==1.2.1
cryptography==1.0.1
Django==1.8.2
django-ratelimit==0.6.0
django-registration-redux==1.2
gunicorn==19.3.0
idna==2.0
kombu==3.0.26
Markdown==2.6.2
paypalrestsdk==1.11.0
psycopg2==2.6
pyasn1==0.1.8
pycparser==2.14
pyOpenSSL==0.15.1
pytz==2015.6
requests==2.7.0
restless==2.0.1
six==1.9.0
